Create PID_Controller_V3
This version includes all functionality but has not been tested with the solenoids plugged in. All temperatures ranges have been modified to lower values for testing purposes. The problem persists that temperature values change when buttons are pressed. This may be due to bad connections due to the temp sensor being attached via breadboard.